requirements.txt
Flask==1.1.2
nltk==3.5
spacy==2.3.2
Flask_Cors==3.0.9
requests==2.24.0
pandas==1.1.4
torch==1.7.1
numpy==1.19.4
transformers==4.6.1
scikit-learn==0.24.2
contractions==0.0.52
python-dotenv==0.17.1
https://github.com/explosion/spacy-models/releases/download/en_core_web_md-2.3.1/en_core_web_md-2.3.1.tar.gz#egg=en_core_web_md==2.3.1